+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/TopActivity.kt
